requirements.txt
# requirements.txt
torch
torchvision
ultralytics
opencv-python
opencv-contrib-python
numpy
faiss-cpu
Pillow 